+++ lib/src/detectors/brightness_analyzer.dart
@@ -65,4 +65,3 @@ class BrightnessAnalyzer {
         maxBrightness: maxBrightness,
       ).classify(averageBrightness);
 }
-